Updated det0/det1, updated TX; gen boot files

+++ MNS_FSW_XQ_preV4_src/src/ReadCommandType.h
@@ -22,7 +22,7 @@ unsigned int GetLastCommandSize( void );
 int ReadCommandType(char * RecvBuffer, XUartPs *Uart_PS);
 int GetIntParam( int param_num );
 float GetFloatParam( int param_num );
-unsigned long long GetRealTimeParam( void );
+unsigned int GetRealTimeParam( void );
 char * GetFileToAccess( void );
 
 #endif /* SRC_READCOMMANDTYPE_H_ */
